refactor: move small fs guard function
